docs: BOJ 15683 G4 감시
